change location of header files for custom MsgTypes
