+++ workspace/custom_msgs/geometry_msgs/msg/quaternion.hpp
@@ -188,7 +188,9 @@ public:
   }
 
   uint32_t getTotalSize(){
-    return cntPub ;
+    uint32_t tmpCntPub = cntPub;
+    cntPub = 0;
+    return tmpCntPub ;
   }
 
 private:
