--- workspace/custom_msgs/geometry_msgs/msg_settings.json
@@ -1,5 +1,0 @@
-{
-    "pubsubMsgs": [
-        "geometry_msgs/msg/Twist.msg"
-    ]
-}
